Change frontend backend communication to JSON

+++ LibraryProject/src/main/java/com/example/libraryproject/servlet/BookKeeperServlet.java
@@ -6,6 +6,7 @@ import com.example.libraryproject.model.dto.BookAdditionRequest;
 import com.example.libraryproject.model.dto.UserDTO;
 import com.example.libraryproject.service.BookKeeperService;
 import com.example.libraryproject.service.GoogleBooksAPIService;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.MultipartConfig;
@@ -24,46 +25,74 @@ import java.util.Set;
 @MultipartConfig
 @WebServlet(name = "BookKeeperServlet", urlPatterns = "/api/bookkeeper/*")
 public class BookKeeperServlet extends HttpServlet {
-
+    private ObjectMapper objectMapper;
 
     @Override
-    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        BookKeeperService bookKeeperService = (BookKeeperService) req.getServletContext()
+    public void init() {
+        objectMapper = (ObjectMapper) getServletContext().getAttribute(ApplicationProperties.OBJECT_MAPPER_ATTRIBUTE_NAME);
+    }
+
+    @Override
+    public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+
+        BookKeeperService bookKeeperService = (BookKeeperService) request.getServletContext()
                 .getAttribute(ApplicationProperties.BOOKKEEPER_SERVICE_ATTRIBUTE_NAME);
-        GoogleBooksAPIService googleBooksAPIService = (GoogleBooksAPIService) req.getServletContext()
+        GoogleBooksAPIService googleBooksAPIService = (GoogleBooksAPIService) request.getServletContext()
                 .getAttribute(ApplicationProperties.GOOGLE_BOOKS_API_ATTRIBUTE_NAME);
 
-        String path = req.getPathInfo();
+        String path = request.getPathInfo();
 
-        switch(path) {
+        switch (path) {
+            case "/upload-image":
+
+                String imageUrl = handleUploadImage(request, bookKeeperService);
+                objectMapper.writeValue(response.getWriter(), imageUrl);
+                response.setStatus(HttpServletResponse.SC_OK);
+                break;
+
             case "/add-book":
-                handleAddBook(req, bookKeeperService);
+                handleAddBook(request, bookKeeperService);
                 break;
+
             case "/add-book-from-google":
-                handleAddBookFromGoogleAPI(req, bookKeeperService, googleBooksAPIService);
+
+                handleAddBookFromGoogleAPI(request, googleBooksAPIService);
                 break;
+
             case "/mark-borrowed":
-                handleMarkBorrowed(req, bookKeeperService);
+
+                handleMarkBorrowed(request, bookKeeperService);
                 break;
 
             case "/ban-user":
-                handleBanUser(req, bookKeeperService);
+
+                handleBanUser(request, bookKeeperService);
                 break;
 
             case "/unban-user":
-                handleUnbanUser(req, bookKeeperService);
+
+                handleUnbanUser(request, bookKeeperService);
                 break;
 
             default:
-                resp.sendError(HttpServletResponse.SC_NOT_FOUND);
+
+                response.sendError(HttpServletResponse.SC_NOT_FOUND);
         }
     }
 
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+
         BookKeeperService bookKeeperService = (BookKeeperService) request.getServletContext()
                 .getAttribute(ApplicationProperties.BOOKKEEPER_SERVICE_ATTRIBUTE_NAME);
         String path = request.getPathInfo();
+
         if ("/users".equals(path)) {
             handleGetUsers(request, response, bookKeeperService);
         } else {
@@ -71,10 +100,14 @@ public class BookKeeperServlet extends HttpServlet {
         }
     }
 
-    private void handleAddBookFromGoogleAPI(HttpServletRequest req, BookKeeperService bookKeeperService, GoogleBooksAPIService googleBooksAPIService) {
-        String title = req.getParameter("title");
-        String author = req.getParameter("author");
-        if(title == null){
+    private void handleAddBookFromGoogleAPI(HttpServletRequest req,
+                                            GoogleBooksAPIService googleBooksAPIService) throws IOException {
+
+        JsonNode params = objectMapper.readTree(req.getReader());
+
+        String title = params.get("title").asText();
+        String author = params.get("author").asText();
+        if (title == null) {
             throw new IllegalArgumentException("Title is required to add a new book.");
         }
         BookAdditionFromGoogleRequest request = new BookAdditionFromGoogleRequest(title, author);
@@ -88,64 +121,76 @@ public class BookKeeperServlet extends HttpServlet {
 
         String path = req.getPathInfo();
 
-        switch (path) {
-            case "/delete-book":
-                handleDeleteBook(req, bookKeeperService);
-                break;
-
-            default:
-                resp.sendError(HttpServletResponse.SC_NOT_FOUND);
+        if (path.equals("/delete-book")) {
+            handleDeleteBook(req, bookKeeperService);
+        } else {
+            resp.sendError(HttpServletResponse.SC_NOT_FOUND);
         }
     }
 
-    private void handleAddBook(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException, ServletException {
-        String title = req.getParameter("title");
-        String author = req.getParameter("author");
-        String description = req.getParameter("description");
-        String genre = req.getParameter("genre");
+    private String handleUploadImage(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException, ServletException {
         Part filePart = req.getPart("image");
-        String imageUrl = "";
-
-        if (filePart != null && filePart.getSize() > 0) {
-            imageUrl = bookKeeperService.downloadImage(filePart, req.getContextPath());
+        if (filePart == null || filePart.getSize() == 0) {
+            throw new IllegalArgumentException("Image file is required.");
         }
+        return bookKeeperService.downloadImage(filePart);
+    }
 
-        if(title == null || author == null){
+    private void handleAddBook(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException {
+
+        BookAdditionRequest request = objectMapper.readValue(req.getInputStream(), BookAdditionRequest.class);
+
+        if (request.author() == null) {
             throw new IllegalArgumentException("Title and author are required to add a new book.");
         }
 
-        BookAdditionRequest bookAdditionRequest = new BookAdditionRequest(title, author, description, genre, imageUrl);
-        bookKeeperService.addBook(bookAdditionRequest);
+        bookKeeperService.addBook(request);
+
     }
 
-    private void handleMarkBorrowed(HttpServletRequest req, BookKeeperService bookKeeperService) {
-        String orderPublicId = req.getParameter("orderPublicId");
-        if(orderPublicId == null){
+    private void handleMarkBorrowed(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException {
+
+        JsonNode params = objectMapper.readTree(req.getReader());
+
+        String orderPublicId = params.get("orderPublicId").asText();
+
+        if (orderPublicId == null) {
             throw new IllegalArgumentException("Order Public ID is required to mark book as borrowed.");
         }
 
         bookKeeperService.tookBook(orderPublicId);
     }
 
-    private void handleDeleteBook(HttpServletRequest req, BookKeeperService bookKeeperService) {
-        String bookPublicId = req.getParameter("bookPublicId");
-        if(bookPublicId == null){
+    private void handleDeleteBook(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException {
+
+        JsonNode params = objectMapper.readTree(req.getReader());
+
+        String bookPublicId = params.get("bookPublicId").asText();
+
+        if (bookPublicId == null) {
             throw new IllegalArgumentException("Order Public ID is required to add a new book.");
         }
 
         bookKeeperService.deleteBook(bookPublicId);
     }
 
-    private void handleBanUser(HttpServletRequest req, BookKeeperService bookKeeperService) {
-        String username = req.getParameter("username");
+    private void handleBanUser(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException {
+
+        JsonNode params = objectMapper.readTree(req.getReader());
+
+        String username = params.get("username").asText();
+
         if (username == null || username.trim().isEmpty()) {
             throw new IllegalArgumentException("Username is required to ban a user.");
         }
         bookKeeperService.banUser(username);
     }
 
-    private void handleUnbanUser(HttpServletRequest req, BookKeeperService bookKeeperService) {
-        String username = req.getParameter("username");
+    private void handleUnbanUser(HttpServletRequest req, BookKeeperService bookKeeperService) throws IOException {
+
+        JsonNode params = objectMapper.readTree(req.getReader());
+
+        String username = params.get("username").asText();
         if (username == null || username.trim().isEmpty()) {
             throw new IllegalArgumentException("Username is required to unban a user.");
         }
@@ -154,7 +199,7 @@ public class BookKeeperServlet extends HttpServlet {
 
     private void handleGetUsers(HttpServletRequest req, HttpServletResponse resp, BookKeeperService bookKeeperService)
             throws IOException {
-        ObjectMapper objectMapper = new ObjectMapper();
+
         Set<UserDTO> users = bookKeeperService.getUsers();
         objectMapper.writeValue(resp.getWriter(), users);
     }
